# openapi-public.json
{"openapi":"3.0.0","paths":{"/api/v1/cards":{"get":{"operationId":"CardApiController_search","summary":"Search cards by name","parameters":[{"name":"ascend","required":false,"in":"query","schema":{}},{"name":"sort","required":false,"in":"query","schema":{}},{"name":"limit","required":false,"in":"query","schema":{}},{"name":"page","required":false,"in":"query","schema":{}},{"name":"q","required":true,"in":"query","description":"Search query","schema":{}}],"responses":{"200":{"description":"Search results"}},"tags":["Cards"]}},"/api/v1/cards/{cardId}/prices":{"get":{"operationId":"CardApiController_getPricesById","summary":"Get current prices for a card by ID","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Card prices"}},"tags":["Cards"]}},"/api/v1/cards/{cardId}/price-history":{"get":{"operationId":"CardApiController_getPriceHistoryById","summary":"Get price history for a card by ID","parameters":[{"name":"cardId","required":true,"in":"path","schema":{"type":"string"}},{"name":"days","required":false,"in":"query","description":"Number of days of history","schema":{"type":"string"}}],"responses":{"200":{"description":"Price history data"}},"tags":["Cards"]}},"/api/v1/cards/{setCode}/{setNumber}/prices":{"get":{"operationId":"CardApiController_getPricesBySetCodeAndNumber","summary":"Get current prices for a card by set code and number","parameters":[{"name":"setCode","required":true,"in":"path","schema":{"type":"string"}},{"name":"setNumber","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Card prices"},"404":{"description":"Card not found"}},"tags":["Cards"]}},"/api/v1/cards/{setCode}/{setNumber}/price-history":{"get":{"operationId":"CardApiController_getPriceHistoryBySetCodeAndNumber","summary":"Get price history for a card by set code and number","parameters":[{"name":"setCode","required":true,"in":"path","schema":{"type":"string"}},{"name":"setNumber","required":true,"in":"path","schema":{"type":"string"}},{"name":"days","required":false,"in":"query","description":"Number of days of history","schema":{"type":"string"}}],"responses":{"200":{"description":"Price history data"},"404":{"description":"Card not found"}},"tags":["Cards"]}},"/api/v1/cards/{setCode}/{setNumber}":{"get":{"operationId":"CardApiController_findBySetCodeAndNumber","summary":"Get card by set code and collector number","parameters":[{"name":"setCode","required":true,"in":"path","schema":{"type":"string"}},{"name":"setNumber","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Card detail"},"404":{"description":"Card not found"}},"tags":["Cards"]}},"/api/v1/sets/{code}/sealed-products":{"get":{"operationId":"SealedProductApiController_findBySet","summary":"List sealed products for a set","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{}},{"name":"page","required":false,"in":"query","schema":{}}],"responses":{"200":{"description":"Sealed products for the set"}},"tags":["Sealed Products"]}},"/api/v1/sealed-products/{uuid}":{"get":{"operationId":"SealedProductApiController_findByUuid","summary":"Get sealed product detail","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Sealed product detail"},"404":{"description":"Not found"}},"tags":["Sealed Products"]}},"/api/v1/sealed-products/{uuid}/price-history":{"get":{"operationId":"SealedProductApiController_getPriceHistory","summary":"Get sealed product price history","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}},{"name":"days","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Price history data"}},"tags":["Sealed Products"]}},"/api/v1/sets":{"get":{"operationId":"SetApiController_findAll","summary":"List sets","parameters":[{"name":"group","required":false,"in":"query","description":"Grouping mode: \"block\" for block-level pagination","schema":{}},{"name":"q","required":false,"in":"query","description":"Search query","schema":{}},{"name":"baseOnly","required":false,"in":"query","description":"Show only base/main sets","schema":{}},{"name":"filter","required":false,"in":"query","description":"Filter sets by name","schema":{}},{"name":"ascend","required":false,"in":"query","schema":{}},{"name":"sort","required":false,"in":"query","schema":{}},{"name":"limit","required":false,"in":"query","schema":{}},{"name":"page","required":false,"in":"query","schema":{}}],"responses":{"200":{"description":"List of sets"}},"tags":["Sets"]}},"/api/v1/sets/{code}":{"get":{"operationId":"SetApiController_findByCode","summary":"Get set detail by code","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Set detail"},"404":{"description":"Set not found"}},"tags":["Sets"]}},"/api/v1/sets/{code}/cards":{"get":{"operationId":"SetApiController_findCardsInSet","summary":"Get cards in a set","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}},{"name":"baseOnly","required":false,"in":"query","description":"Show only base set cards","schema":{}},{"name":"filter","required":false,"in":"query","description":"Filter cards by name","schema":{}},{"name":"ascend","required":false,"in":"query","schema":{}},{"name":"sort","required":false,"in":"query","schema":{}},{"name":"limit","required":false,"in":"query","schema":{}},{"name":"page","required":false,"in":"query","schema":{}}],"responses":{"200":{"description":"Cards in set"}},"tags":["Sets"]}},"/api/v1/sets/{code}/price-history":{"get":{"operationId":"SetApiController_getPriceHistory","summary":"Get set price history","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}},{"name":"days","required":false,"in":"query","description":"Number of days of history","schema":{"type":"string"}}],"responses":{"200":{"description":"Set price history data"}},"tags":["Sets"]}}},"info":{"title":"I Want My MTG — Public API","description":"Read-only catalog endpoints (cards, sets, sealed products, prices, price history). No authentication required when called via the RapidAPI marketplace.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateApiKeyDto":{"type":"object","properties":{"name":{"type":"string","description":"User-supplied label for the key","maxLength":64}},"required":["name"]},"LoginRequestDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string"}},"required":["email","password"]},"CheckoutRequestDto":{"type":"object","properties":{"plan":{"type":"string","enum":["monthly","annual"]}},"required":["plan"]},"CreatePriceAlertDto":{"type":"object","properties":{}},"UpdatePriceAlertDto":{"type":"object","properties":{}},"SealedInventoryRequestDto":{"type":"object","properties":{"sealedProductUuid":{"type":"string"},"quantity":{"type":"number"}},"required":["sealedProductUuid","quantity"]},"SealedInventoryDeleteRequestDto":{"type":"object","properties":{"sealedProductUuid":{"type":"string"}},"required":["sealedProductUuid"]},"TransactionRequestDto":{"type":"object","properties":{}},"TransactionUpdateRequestDto":{"type":"object","properties":{}},"UpdateUserRequestDto":{"type":"object","properties":{}},"UpdatePasswordRequestDto":{"type":"object","properties":{}}}}}